+++ src/styles/Task.css
@@ -24,3 +24,9 @@
   padding: 0;
   width: 100%;
 }
+
+#deleteButton {
+  padding-bottom: 0;
+  padding-top: 0;
+  height: 100%;
+}
